fix decrease thirsty

+++ Howling/Assets/Scripts/UI/StatusController.cs
@@ -128,7 +128,7 @@ public class StatusController : MonoBehaviour
                 currentThirstyDecreaseTime++;
             else
             {
-                currentThirsty -= (int)(Time.deltaTime * 40f);
+                currentThirsty -= Time.deltaTime * 40f;
                 currentThirstyDecreaseTime = 0;
             }
         }
